feat(adr-061): split source-graph values out of buildsource/source_graph.py
Starts ADR-061 Phase 5 item 2 ("separate source-graph values,
construction, and comparison"). Moves the values third into
abicheck/model/source_graph.py: SourceGraphSummary (ADR-031 D7's compact
graph container, with all its mutation/serialization methods),
GraphSummaryDiff, the node-id constructors (_source_node_id and its ten
siblings, function_decl_identity), and the schema vocabulary
(NODE_KINDS/EDGE_KINDS/DEPENDENCY_EDGE_KINDS/SOURCE_GRAPH_VERSION/
EVIDENCE_TIER_L5).

buildsource/source_graph.py imports and re-exports every moved name
(X as X, matching the file's own pre-existing graph_facts.py re-export
convention for mypy's --no-implicit-reexport), so all 77 existing
callers keep resolving. File drops from 2000 to 1352 lines; its
architecture/debt.yaml no-growth baseline is lowered to match.

_conf_from_build stayed behind: it needs build_evidence.py's Confidence
enum, and build_evidence.py transitively imports the extract-classified
comdat_groups.py — classifying build_evidence.py model to satisfy that
one function would create a real model -> extract cycle, caught by
check_architecture.py before this landed. entity_resolver.py had no such
coupling, so it's now classified model in architecture/modules.yaml
(virtual classification, still physically flat).

Construction (build_source_graph) and comparison (diff_source_graph)
remain in buildsource/source_graph.py — a separate, not-yet-attempted
follow-up slice, same as the first ADR-061 Phase 5 item 1 slice's
approach to dumper_castxml.py.

Verified: check_architecture.py 0 errors, mypy clean, ruff clean,
tests/test_source_graph.py + test_consumer_graph.py + test_use_cases.py
(210 tests, the ones referencing NODE_KINDS/EDGE_KINDS through the flat
re-export) pass unchanged.

Updates the ADR's status line and item-2 entry to record this honestly
as started, not done.
